implemented department related components

+++ EMMS-Backend/src/main/java/com/emms/dao/EquipmentRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.emms.model.Equipment;
 
-public interface EquipmentRepository extends JpaRepository<Equipment, Integer> {
+public interface EquipmentRepository extends JpaRepository<Equipment, Long> {
 
 }
